Moved test files.
