+++ test/db/Edge.js
@@ -1,4 +1,4 @@
-const db = require('../app/db');
+const db = require('../../app/db');
 const assert = require('assert');
 
 const {
